Do not treat empty directory as key store v1
There is no definite easy way to tell whether a directory is key store
v1 because it does not have any identifying features like key store v2
has: no version files, nothing. So we have used a conservative approach:
if it's a directory, it's a key store v1.

However, this does not account for the case of the empty directory. Yes,
it is a directory, but it's probably not a *key* directory. This may
cause tools -- which all use autodetetion now -- falsely believe that
a directory *is* a key store v1. In particular, acra-keymaker will think
that it's a key store v1. It will not require an explicit "--keystore"
argument, and will generate v1 keys, which may not be what the user
wants to achieve (or rather, what we would like the user to do).

The users might first create a directory, then pass it to "--keys_dir"
argument, forget the "--keystore" argument, and they will silently get
key store v1. Make sure that they will get an error instead, prompting
an explicit choice of the version with "--keystore".