Detect directory only when not specified for acra-keymaker
acra-keymaker incorrectly runs key store version detection when the
version is explicitly requested by the user. This might trick it into
falsely believing that it should create key store v1 when the users
asks to create a key store v2.

acra-keymaker should detect the version only when the user did not
specify it explicitly. That's how Acra EE does it. Apparently, this
condition was lost when the code was ported from Acra EE to Acra CE.

This particular bug has caused a very contiving sequence of events
leading to failures in some Acre EE tests when key store v2 is used
because some of the generated keys were generated in v1 format and
then expected to be read in v2 format.