Reusable CommonKeyStoreParameters
All (or almost all) subcommands of "acra-keys" are going to need key
store configuration parameters. Those parameters are all the same
so they can be made into a reusable block.

In most cases the parameters do not need prefixes, but sometimes they
are necessary when a command involves multiple key stores.