Don't handle --dump_config immediately in ParseFlagsWithConfig
Instead of dumping configuration for current flag set and exiting
immediately, return a special error from ParseFlagsWithConfig(),
like the standard Go parser does with the "--help" option.

However, keep the behavior of Parse() the same as before. Multiple
utilities rely on this.

The only user of ParseFlagsWithConfig() is "acra-keys" where this change
does not affect the observable behavior.