Decouple acra-keys subcommands (#396)
* Reusable CommonKeyStoreParameters

All (or almost all) subcommands of "acra-keys" are going to need key
store configuration parameters. Those parameters are all the same
so they can be made into a reusable block.

In most cases the parameters do not need prefixes, but sometimes they
are necessary when a command involves multiple key stores.

* Parameters for "acra-keys list"

This struct holds all parameters for "acra-keys list" subcommand. Key
list formatting parameters are used by "acra-keys import" as well so
they are split into another common mix-in struct.

* Parameters for "acra-keys export" and "import"

There are some common parameters for those subcommands as well, those
are reused as much as possible.

The exportKeyIDs parameter is a positional one, not an option. There is
a field for it to satisfy the interface, but we'll deal with it later.

* Parameters for "acra-keys read"

No common parameters now, just a single positional parameter for the key
kind to read and a couple of options, along with the common parameters.

* Parameters for "acra-keys destroy"

These are similar to "acra-keys read", but we don't need zone ID because
only transport keys can be destroyed.

* Config-enabled flag parsing for any flag set

Introduce ParseFlagsWithConfig so that the same config/flag parsing
logic can be used with any flag set, not just the global one.

* Parse "acra-keys list" arguments

This subcommand is easy: it has only flags, no positional arguments, and
all flags are optional. We can just rely on the common parsing routines.

* Parse "acra-keys export" arguments

This subcommand is more complex. There is some (common) validation logic
for required options specifying the input/output files. Additionally,
"acra-keys export" has to deal with positional arguments with key IDs
to export.

* Parse "acra-keys import" arguments

Importing is easier now since it's all common parsing and validation.
There are no positional parameters either.

* Parse "acra-keys read" arguments

This subcommand has exactly one required positional argument. The reason
it's not an option is that we'd like to reserve the ability to accept
mulitple key kinds in the future.

There is also some custom validation logic for required flag sets
depending on the key kind.

* Parse "acra-keys destroy" arguments

This subcommand is similar to "acra-keys read", just doesn't need zone IDs.

* New subcommand parameter parsing

Introduce a new common Subcommand interface which organizes available
subcommands and enables common handling of them. Now the set of the
subcommands can vary, and different implementations of subcommands can
be used.

Note that the new parsing code does not have any global options defined
by subcommands. It accepts only generic options related to config files,
configuration dumping, command-line help, etc.

In particular, the key store configuration has to specified *after* the
subcommand name, not before. This call:

    acra-keys --keys_dir my-custom-key-store list

should now be written like this:

    acra-keys list --keys_dir my-custom-key-store

which is arguably more predictable and consistent.

* Drop old command-line parsing code

The old parsing code is now no longer used and can be removed.

* Move subcommand execution into Subcommand

Move all details of subcommand execution into the Subcommand interface,
using a new method Execute(). Now we can get rid of this giant switch,
dispatching by names, and weird casts.

* Update argument order in tests

Now "acra-keys" accepts "--keys_dir" and other arguments after the
subcommand and does not need this parameter for configuration output.