Full config for acra-keys (#397)
* Don't handle --dump_config immediately in ParseFlagsWithConfig

Instead of dumping configuration for current flag set and exiting
immediately, return a special error from ParseFlagsWithConfig(),
like the standard Go parser does with the "--help" option.

However, keep the behavior of Parse() the same as before. Multiple
utilities rely on this.

The only user of ParseFlagsWithConfig() is "acra-keys" where this change
does not affect the observable behavior.

* More flexible DumpConfigFromFlagSets

The current DumpConfig() function is too tightly coupled with global
command-line flags from "flag.CommandLine". It iterates only over the
global flags. This is okay for most utilities as they use only the
global flag set. However, "acra-keys" has subcommands, each of which
has its own specific flag set in addition to the global one.

Introduce a set of functions which accept a list of flag sets to act
upon, and reimplement existing functions with them. There are many users
of existing functions so we keep their behavior the same: use only the
flag.CommandLine flag set.

Several flag sets can be used because "acra-keys" uses both the global
flag set and several flag sets for subcommands. Furthermore, the flag
sets often overlap, so we deduplicate configuration options by name.

* Flag set accessors for subcommands

We'll need to access the FlagSet fields of each subcommand so add a
method for that.

* Dump all flag sets of acra-keys

Instead of dumping only the global flag set, use all available
subcommands when "--dump_config" is on the command line.