Common ColumnDecryptionNotifier implementation
Keeping a list of DecryptionSubscriber and notifying them about column
decryption events is a more or less independent piece of functionality
which can be shared between PostgreSQL and MySQL decryptors.

Introduce a shared ColumnDecryptionObserver which provides
ColumnDecryptionNotifier interface and an entry point for the proxies
to trigger notifications.

It is based on current PostgreSQL implementation, but uses slices
instead of linked list for better performance of notifications which are
going to be more common than subscriber registration. The implementation
also does some primitive deduplication so that a single subscriber
cannot be accidentally subscribed multiple times for the same event.