Observe decryption events for text data rows
Insert onColumnDecryption() notifications when processing text data
rows. This is easy for the most part: process the data before we try
decrypting it. Note that this process may change the "value" so we
should take care to insert the possibly updated value into the output
rather than copy the original data if the field is not suitable for
decryption. Some fields may still be processed with other goals.