Observe decryption events for binary rows: decryptable strings
There are many binary data formats in MySQL. The easiest one to handle
is the format for binary strings. These are decryptable data formats
like varchar or blob. Use the value as it and account for possible
update of the value after processing, taking care to reencode and
write it back.