Observe decryption events for binary rows: fixed-size numbers
These fields are tricky. Binary format of MySQL stores them compactly,
but subscribers expect to work on strings. Therefore we need to decode,
format, process, parse, and then encode the results back into the same
binary format as they had. This involves quite a lot of copypasta for
type-switching but it is what it is.