Require --keystore version when generating master keys
Previously the "--generate_master_key" option did not require the user
to specify the key store version. It always outputs a single symmetric
key, and for key store v2 the user was supposed to generate two keys
separately.

However, this is going to change. "acra-keymaker" will output a suitable
master key in one call, even for key store v2. For this it will need to
know which format to generate.

Make the "--keystore" option required when using "--generate_master_key"
even if a master key for key store v1 is requested. This is a breaking
change because from now on it will not be possible to use

    acra-keymaker --generate_master_key path/to/master.key

for key store v1 keys.