Common master key format for key store v2
"acra-keys export" already suggests a new master key format for key
store v2: a JSON object with two keys inside. This format is more bulky
than a simple binary format, but it is way more future-proof.

Move the ad-hoc implementation of this format into the shared module.
It is going to be reused for master key ingestion in other tools.