Use different variables for acra-migrate-keys
acra-migrate-keys deals with two separate key stores which may have
different master keys. Instead of accepting master keys via the same
ACRA_MASTER_KEY variable, use distinct {SRC,DST}_ACRA_MASTER_KEY,
similar to how the command-line options use the same prefix.

Testing this is a bit move involved since previously migration tests
expected that all variables are already configured and unchaged.
Now this is not the case and we have to pass them explicitly, after
generating a new set of master keys.

In order to set environment variables, we need to control the arguments
passed to popen() call. fork_acra() method already has it in the form of
optional "popen_kwargs" argument. Add the same thing to fork_connector()
to be able to control the environment of launched AcraConnector.

In other places we don't bother because threading the environment vars
is too bothersome to implement.