Unify errors in case of invalid master key settings
Take this chance to unify error messages printed when master key
settings are incorrect: either the environment variable is not set,
or it is set to an invalid value.

Make sure to include necessary context, such as the environment variable
name (important for key migration). However, don't be too specific
because there are other options for master key storage (in Acra EE).