Generating master keys
Add support for generating master keys when "--master_key_path" is
specified. Master keys are special since that's the first thing which
needs to be generated before "acra-keys generate" can generate other
keys in the keystore.

The user is required to specify the keystore version for which to
generate a new master key.