Use single ACRA_MASTER_KEY for key store v2 (#399)
* Require --keystore version when generating master keys

Previously the "--generate_master_key" option did not require the user
to specify the key store version. It always outputs a single symmetric
key, and for key store v2 the user was supposed to generate two keys
separately.

However, this is going to change. "acra-keymaker" will output a suitable
master key in one call, even for key store v2. For this it will need to
know which format to generate.

Make the "--keystore" option required when using "--generate_master_key"
even if a master key for key store v1 is requested. This is a breaking
change because from now on it will not be possible to use

    acra-keymaker --generate_master_key path/to/master.key

for key store v1 keys.

* Common master key format for key store v2

"acra-keys export" already suggests a new master key format for key
store v2: a JSON object with two keys inside. This format is more bulky
than a simple binary format, but it is way more future-proof.

Move the ad-hoc implementation of this format into the shared module.
It is going to be reused for master key ingestion in other tools.

* Generate and read new key store v2 master keys

Key store v2 requires two separate master keys: one for encryption of
key material and another one for signing the key store metadata. The
original idea was to naturally use two separate environment variables
to supply the key values. While it is easy to code, it is not really
easy for the users to use. They have to generate two separate keys,
store them separately, set them up separately, while conceptually
they are still just one "Acra master key".

So... Instead of using two separate variables for key store v2, use the
same ACRA_MASTER_KEY as key store v1 uses. However, instead of expecting
base64-encoded key bytes there, we expect a base64-encoded JSON object
with base64-encoded keys in it. This is the key format which can be
generated with

    acra-keymaker --keystore=v2 --generate_master_key master.key

in one go, and then read in using

    export ACRA_MASTER_KEY=$(cat master.key | base64)

While double-wrapping base64 may sound silly and wasteful, it simplifies
setup flow for the users: always pass the key file to base64 before
setting the ACRA_MASTER_KEY variable. There is no need to special case
key store v2 just because it's already readable JSON inside.

For example, with key store v1 the master is simply 32 random bytes,
encoded in base64 for the environment:

    5Tl/mehY/QhswVMn13QUvSu9MdSsm+JvRefWGF/zBsA=

while for key store v2 the key consists of two sets of 32 random bytes,
wrapped into a JSON object:

    {
      "encryption":"dN6qtsv+yi1YeCCnKrHuGyn9vOdLVTkAd4/p7pfYIVk=",
      "signature":"lI4ziBwzWGIz5z7JUsKZgPfPeBHCquaxOaBl5ZMJnEo="
    }

which is then encoded in base64 again for ACRA_MASTER_KEY variable:

    eyJlbmNyeXB0aW9uIjoiZE42cXRzdit5aTFZZUNDbktySHVHeW45dk9kTFZUa0FkNC9wN3BmWUlWaz0iLCJzaWduYXR1cmUiOiJsSTR6aUJ3eldHSXo1ejdKVXNLWmdQZlBlQkhDcXVheE9hQmw1Wk1KbkVvPSJ9

While admittedly wasteful, this greatly simplifies the setup flow:

    acra-keymaker --keystore=v1 --generate_master_key master-key.v1
    export ACRA_MASTER_KEY=$(cat master-key.v1 | base64)

or s/v1/v2/g for key store v2.

* Use unified master key in test suite

Instead of generating multiple master key, get back to using a single
master key. However, take care to fix up tests which verify the format
of the key. There we need to format a proper JSON object for the key.

* Use different variables for acra-migrate-keys

acra-migrate-keys deals with two separate key stores which may have
different master keys. Instead of accepting master keys via the same
ACRA_MASTER_KEY variable, use distinct {SRC,DST}_ACRA_MASTER_KEY,
similar to how the command-line options use the same prefix.

Testing this is a bit move involved since previously migration tests
expected that all variables are already configured and unchanged.
Now this is not the case and we have to pass them explicitly, after
generating a new set of master keys.

In order to set environment variables, we need to control the arguments
passed to popen() call. fork_acra() method already has it in the form of
optional "popen_kwargs" argument. Add the same thing to fork_connector()
to be able to control the environment of launched AcraConnector.

In other places we don't bother because threading the environment vars
is too bothersome to implement.

* COMMENCING COMMEMORATIVE COMMA COMMIT

* Unify errors in case of invalid master key settings

Take this chance to unify error messages printed when master key
settings are incorrect: either the environment variable is not set,
or it is set to an invalid value.

Make sure to include necessary context, such as the environment variable
name (important for key migration). However, don't be too specific
because there are other options for master key storage (in Acra EE).