Use "keystore" in error messages and logs
Update source code to spell "keystore" as a single word. This is
preferable since that's how Go modules are called, how the command-line
operations are named, and in general to convey the unitary meaning
without sounding like a place where they sell keys.

Note that the code also has a lot of CamelCase KeyStore which suggests
separate words, but we don't change that to avoid breaking stuff. Only
human-visible strings are changed.