Spell "keystore" as a single word (#401)
* Use "keystore" in error messages and logs

Update source code to spell "keystore" as a single word. This is
preferable since that's how Go modules are called, how the command-line
operations are named, and in general to convey the unitary meaning
without sounding like a place where they sell keys.

Note that the code also has a lot of CamelCase KeyStore which suggests
separate words, but we don't change that to avoid breaking stuff. Only
human-visible strings are changed.

* Use "keystore" in comments

Update comments as well to avoid suggesting that this is alternate
spelling. Otherwise developers may read a lot of comments and write
new code using the undesirable spelling.

* Use "keystore" in CircleCI tasks

* Use "keystore" in version string

This change is somewhat important since it changes key store format.
However, we can still make this change for consistency as key store v2
has not yet been available to the users.

* Update configuration files