MySQL decryption notifications (#398)
* Integer support in MysqlDBDataCoder

Add support for sqlparser.IntVal in MysqlDBDataCoder. This allow various
parts of code to understand SQL queries which use integers in them. The
integers are represented with strings, exactly as in the query.

Update tests as well to expect the new behavior and clean them up a bit.

* Common ColumnDecryptionNotifier implementation

Keeping a list of DecryptionSubscriber and notifying them about column
decryption events is a more or less independent piece of functionality
which can be shared between PostgreSQL and MySQL decryptors.

Introduce a shared ColumnDecryptionObserver which provides
ColumnDecryptionNotifier interface and an entry point for the proxies
to trigger notifications.

It is based on current PostgreSQL implementation, but uses slices
instead of linked list for better performance of notifications which are
going to be more common than subscriber registration. The implementation
also does some primitive deduplication so that a single subscriber
cannot be accidentally subscribed multiple times for the same event.

* Use common observer in PostgreSQL decryptor

Replace the current ad-hoc implementation with ColumnDecryptionObserver.

* Implement ColumnDecryptionNotifier for MySQL proxy

Initialize an observer, export approriate interface, and prepare to
issues MySQL-specific notifications. The new onColumnDecryption() method
will be used in upcoming commits.

* Observe decryption events for text data rows

Insert onColumnDecryption() notifications when processing text data
rows. This is easy for the most part: process the data before we try
decrypting it. Note that this process may change the "value" so we
should take care to insert the possibly updated value into the output
rather than copy the original data if the field is not suitable for
decryption. Some fields may still be processed with other goals.

* Observe decryption events for binary rows: decryptable strings

There are many binary data formats in MySQL. The easiest one to handle
is the format for binary strings. These are decryptable data formats
like varchar or blob. Use the value as it and account for possible
update of the value after processing, taking care to reencode and
write it back.

* Observe decryption events for binary rows: other strings

There are other string-like types which are not supported for
encryption. Process them similarly to decryptable strings.

Exact content of these values may have different internal format, but
they all have a high-level format of a length-prefixed byte string.

* Observe decryption events for binary rows: fixed-size numbers

These fields are tricky. Binary format of MySQL stores them compactly,
but subscribers expect to work on strings. Therefore we need to decode,
format, process, parse, and then encode the results back into the same
binary format as they had. This involves quite a lot of copypasta for
type-switching but it is what it is.

* Require DecryptionSubscribers to ID themselves

Add a required method for self-identification to DecryptionSubscriber
interface. This is useful for error reporting as it allows to pinpoint
which subscriber has failed processing.

Absence of this method also prevents Acra EE from reusing Acra CE
interfaces. Adding is here will allow to remove some duplication.

* Always update data after processing

Do this not only if the data size has been reduced (as an apparent
effect of AcraStruct decryption), but even if it is unchanged or even
increased. Some processing flows (e.g., in Acra EE features such as
pseudonimization) may change the data without chaging its size.

* Special handling of NULL value encoding

As described in MySQL format documentation [1], NULL values are encoded
with a special form: [0xFB] which looks like zero-length string.

[1]: https://dev.mysql.com/doc/internals/en/string.html#packet-Protocol::LengthEncodedString

LengthEncodedString() correctly parses such string and indicates this
with its "isNull" return value. But interestringly, not a single user of
this function uses that value, ignoring the NULL values and treating
them as empty strings.

This was not important before because the implementation did not process
NULL values and copied them from the original packet data. However, now
NULL values can be processed and we no longer can blindly copy them from
the packet. Instead, the new code expects NULL values to be correctly
encoded by PutLengthEncodedString(). But PutLengthEncodedString() did not
encode NULL values correctly. Instead, it treated them as empty strings
which breaks MySQL packets.

In order to resolve this issue, represent NULL values as "nil" in Go. That
way LengthEncodedString() can return "nil", processing can see this "nil",
and PutLengthEncodedString() can encode "nil" into [0xFB] as required.

Simplify LengthEncodedString() interface by removing the addtional
unused return value and consistently return "nil" only when faced with
NULL values. Use an empty slice instead of "nil" if the data is actually
an empty string

We already have a test verifying empty an NULL handling (TestEmptyValues).

* Propagate DataProcessor in mysql.Decryptor

MySQL decryptor is based on PostgreSQL decryptor internally. It does not
override all methods that use data processors, so some of them use the
base implementation. Therefore, SetDataProcessor() needs to update the
data processor for both MySQL methods and the base ones as well.
Otheriwise calling them might cause nil dereferences.

* Decrypt AcraStructs in MySQL via OnColumn callbacks

Instead of hardwiring decryption into processing, sign up the decryptor
for OnColumn notifications, like PostgreSQL code does. This makes the
code nicely unified.

Common decryptor will take care to figure out whether a data field can
contain AcraStructs. It will also bump the performance counters, add
attributes, etc.