Make TableSchema an interface
Avoid exposing bare structures in in interfaces since this makes the
interfaces impossible to extend in Acra EE.

With that in mind, reclassify TableSchema as interface instead of a
struct, and update the code accordingly. Provide accessors instead
of direct field references. Add new methods. Update use sites.

Some fields have to be renamed because in Go you cannot have a field and
a method with the same name. It is also not possible to make the fields
unexported as they need to be exported for YAML parser to access them
via reflection.