Make ColumnEncryptionSettings an interface
In a similar vein, reclassify ColumnEncryptionSettings as an interface.
In fact, we already have EncryptionSettings in the parent package, just
move it into "encryptor/config" and extend with more methods.

Note, however, that the struct is still available as
BasicColumnEncryptionSettings so that it can be reused in Acra EE.

Similar to TableSchema, some exported fields need to be renamed to avoid
conflicts with method names.

Changing return type of "GetColumnEncryptionSettings" from a pointer to
an interface has implications for callers which compare the result with
"nil". In Go a nil interface is *not* equal to nil pointer cast to the
interface type. Therefore, we need to check for nil explicitly and
return a nil interface when appropriate. (I have spent three hours
debugging this issue. Please don't break it again.) Curiously enough,
"GetTableSchema" already uses the 'correct' pattern.