TEST_CLEAN_BINARIES switch to avoid cleanup
Add a switch CLEAN_BINARIES which can selectively disable cleaning up
the binaries used for testing. This can be useful for repeatedly running
integration tests locally, to save up the time spent on rebuilding. Set
the environment variable TEST_CLEAN_BINARIES=0 to disable cleanup and
reuse the same binaries over and over (they are still refreshed in the
code changes).

Additionally, add another variable BUILD_BINARIES which can be used to
disable building the binaries on startup. This will be useful to Acra EE
test suite which would like to provide its own binaries instead. Since
this feature is not useful to developers there is no environment switch.