Do not let acra-heartbeat to launch in config tests
Launching "acra-heartbeat" from Acra EE with default configuration and
no command-line arguments successfully starts the service. This is not
what we want, so tweak the command-line defaults to fail execution if
there are no issues with the configuration file. This is similar to what
we do for AcraServer and other services.