Utility for reading boolean values from environment
Currently, TEST_MYSQL variable is read from environment using bool(...)
conversion which is incorrect. bool() returns True for anything which is
is not an empty string, so TEST_MYSQL=0 or TEST_MYSQL=no would still
trigger MySQL testing. In fact, the current code looks whether the
variable is set, disregarding its value.

Introduce a utility function which reads boolean values from environment
in a more human-friendly way. Now "yes", "y", "true", "1" all enable
MySQL testing, while "no", "n", "false", and "0" disable it. If the
variable is not set or empty, MySQL is not tested (as before).