Test harness improvements (#402)
* Default service handling in config tests

Handle the case when "service" is not present in "default_args" list.
In that case "test_data" should be a list (of arguments) to be appended,
not the default None value returned by "get".

* TEST_CLEAN_BINARIES switch to avoid cleanup

Add a switch CLEAN_BINARIES which can selectively disable cleaning up
the binaries used for testing. This can be useful for repeatedly running
integration tests locally, to save up the time spent on rebuilding. Set
the environment variable TEST_CLEAN_BINARIES=0 to disable cleanup and
reuse the same binaries over and over (they are still refreshed in the
code changes).

Additionally, add another variable BUILD_BINARIES which can be used to
disable building the binaries on startup. This will be useful to Acra EE
test suite which would like to provide its own binaries instead. Since
this feature is not useful to developers there is no environment switch.

* Run keymaker tests in separate directories

Some keys (like poison record keys)  are generated lazily, on first
request, and then left unchanged on future generation requests. This
means that repeated calls to acra-keymaker must be performed with the
same master key. If the master key changes, keymaker will not be able
to read the saved poison record key.

Run the key length tests in separate temporary key directories to make
sure that we're using each master key only once.

* Do not let acra-heartbeat to launch in config tests

Launching "acra-heartbeat" from Acra EE with default configuration and
no command-line arguments successfully starts the service. This is not
what we want, so tweak the command-line defaults to fail execution if
there are no issues with the configuration file. This is similar to what
we do for AcraServer and other services.

* Improve config regeneration

Update the script to regenerate configuration files for all binaries,
now you don't need to add new binaries here for them to be hooked up
for testing (tests already list all tools in the "cmd" directory).

Additionally, check for preexisting binaries in the current directory
which can be used to generate configuration files. Tests run this script
multiple times and they already have all the binaries ready, there is no
need to compile them with "go run" every time, shaving off a couple of
seconds from the build time.

* Explain cached binary usage in "regenerate.sh"

Leave a comment there so that people would not want to "optimize" this
part of the script.

* Utility for reading boolean values from environment

Currently, TEST_MYSQL variable is read from environment using bool(...)
conversion which is incorrect. bool() returns True for anything which is
is not an empty string, so TEST_MYSQL=0 or TEST_MYSQL=no would still
trigger MySQL testing. In fact, the current code looks whether the
variable is set, disregarding its value.

Introduce a utility function which reads boolean values from environment
in a more human-friendly way. Now "yes", "y", "true", "1" all enable
MySQL testing, while "no", "n", "false", and "0" disable it. If the
variable is not set or empty, MySQL is not tested (as before).

* Add "utils" to import list

This allows to refer to utilites as "utils.<...>" in case you don't want
to import it into global scope using that ever-growing list.

* Use new utility for TEST_CLEAN_BINARIES as well

* Ensure "acra-heartbeat" uses plaintext format

Due to hysterical raisins, this tool has been using JSON logging format
by default while other tools use plaintext. Configuration tests expect
tools to output logs in plaintext format to match the error messages,
so make sure that acra-heartbeat uses plaintext.