Generating zone keys
In addition to "acra-keymaker" keys, teach "acra-keys generate" to work
with zone storage keys. Now it can stand in for "acra-addzone" with the
"--new_zone" option. It is also possible to rotate an existing zone
storage key with "--zone_storage_key" and "--zone_id" options (there
were no tool for this before).