Use key IDs in "acra-keys read"
Update this subcommand CLI to accept the same key IDs as "export"
subcommand, and as shown by the "list" subcommand. This unifies the
interface of "acra-keys".

Since key IDs do not indicate whether the public or private key is
requested to be read, introduce a couple of new options for the
"acra-keys read" subcommand to indicate which key should be read.

This subcommand is mostly used in integration tests. Update the
utilities to use the new syntax.