Compare imported string list without order
Go 1.15 has changed serialization of "SET OF" elements in ASN.1 [1]
to be sorted. This means that now the serialized form of exported key
rings does not preserve the order in which they have been given to the
Export() method. Therefore, we cannot depend on Import() to return their
names after import in the same order. Replace the slice comparison with
unordered comparison.

[1]: https://golang.org/doc/go1.15#encoding/asn1