Fix Go 1.15 regression in keystore tests (#412)
* Compare imported string list without order

Go 1.15 has changed serialization of "SET OF" elements in ASN.1 [1]
to be sorted. This means that now the serialized form of exported key
rings does not preserve the order in which they have been given to the
Export() method. Therefore, we cannot depend on Import() to return their
names after import in the same order. Replace the slice comparison with
unordered comparison.

[1]: https://golang.org/doc/go1.15#encoding/asn1

* Avoid depending on ordering in tests

Some tests (like duplication) has depended on keys being exported and
imported sequentially in the given order. This is no longer the case
with Go 1.15 so tweak the tests to avoid such dependency.