Unify key ID paths in "acra-keys" (#411)
* Key ID parser in "acra-keys"

Some commands like "acra-keys read" and "acra-keys destroy" do not
forward key ID strings to keystore as the interfaces they use do not
support key IDs directory. Introduce a helper function "ParseKeyKind"
which parses a key ID into a key kind identifier used in "acra-keys".

* Use key IDs in "acra-keys read"

Update this subcommand CLI to accept the same key IDs as "export"
subcommand, and as shown by the "list" subcommand. This unifies the
interface of "acra-keys".

Since key IDs do not indicate whether the public or private key is
requested to be read, introduce a couple of new options for the
"acra-keys read" subcommand to indicate which key should be read.

This subcommand is mostly used in integration tests. Update the
utilities to use the new syntax.

* Use key IDs in "acra-keys destroy"

Similarly, update the "acra-keys destroy" subcommand to accept key IDs.
This subcommand is a bit simpler since it currently supports only
transport keypairs and removes them completely. It is also mostly useful
for integration testing right now.