Do not validate "client ID" in keystore v1
Remove this check as it interferes with poison record keys. The "id"
parameter is not *client ID* specifically but can also be zone ID,
as well as some other hardcoded context IDs (like poison record one).
Poison records use the path to poison record key as ID, which is not
a valid client ID.

It is not really necessary to validate context ID during key decryption.
If the ID is not correct then the key will simply fail to decrypt.
Validity of actual client IDs is checked during key generation.
Therefore we can safely omit this check here.