Use historical poison record keys (#413)
* Access historical poison private keys

Currently it is not possible to rotate poison record keys, but let's get
ready for that and unify AcraStruct handling across Acra codebase,
always using all rotated keys to decrypt them.

First of all, we need methods to access historical private keys used
to create poison records. Introduce a new keystore method for that.

* Decrypt poison records with all keys

Use all available private keys to try and decrypt poison records.
This makes poison record handling future-proof when it will be possible
to rotate poison record keys.

* Handle missing keys in rotated decryption

DecryptRotatedAcrastruct() did not return an error if it was given
an empty list of private keys to try and, naturally, did not decrypt
anything. Return a special error in this case.

* Do not validate "client ID" in keystore v1

Remove this check as it interferes with poison record keys. The "id"
parameter is not *client ID* specifically but can also be zone ID,
as well as some other hardcoded context IDs (like poison record one).
Poison records use the path to poison record key as ID, which is not
a valid client ID.

It is not really necessary to validate context ID during key decryption.
If the ID is not correct then the key will simply fail to decrypt.
Validity of actual client IDs is checked during key generation.
Therefore we can safely omit this check here.

* Add missing private key zeroization

Now all poison record checks will clean up the private keys after use.